removed more redundant statsample functions
